refactor: move UserService to infrastructure package
The UserService is coupled to the Spring implementation and should not
be in the domain.
